+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.56 - progress]
+### Added
+- `GeoDataExtra.PAR_LOPOINT_REVIEWS` as container for LoPoint reviews
+- `Location.gnssLastCorrectionTime` parameter
+
 ## [0.9.55]
 ### Fixed
 - [#54](https://github.com/asamm/locus-api/issues/54), missing "Virtual" cache size implementation